requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
streamlit
langchain_groq
langchain_core
Fastapi
uvicorn
langserve[all]
httpx
langchain_community
wikipedia
duckduckgo-search
PyPDF2
huggingface_hub
crewai
crewai_tools
ollama
google.generativeai
pinecone-client
pinecone-text
pinecone-notebooks
neo4j==5.25.0
langchain_neo4j
langgraph
langsmith
cassio
langchain-chroma
llama_index
